- moving deterministic losses to determinism - adding printing for adaptive avg pooling deterministic replacement
